event-collector/requirements.txt
fastapi
uvicorn
alembic
pytest-asyncio
redis
psycopg2-binary
pydantic
pydantic-settings
asyncpg
sqlalchemy
httpx
